sentry - platform - PlatformApi - Move from sentry - routes - units-for-slot - Application - add PlatformApi - Fix all calls of Application::new()
