+++ sentry/src/main.rs
@@ -10,7 +10,7 @@ use primitives::{
 };
 use sentry::{
     db::{postgres_connection, redis_connection, setup_migrations, CampaignRemaining},
-    Application,
+    Application, platform::PlatformApi,
 };
 use slog::info;
 use std::{env, net::SocketAddr};
@@ -106,6 +106,11 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
 
     let campaign_remaining = CampaignRemaining::new(redis.clone());
 
+    // todo: Make platform_url configurable! Load from config or pass with env. variable
+    let platform_url = "https://platform.adex.network".parse().expect("Bad ApiUrl, load from Config?");
+    // todo: Make keep_alive_interval configurable!
+    let platform_api = PlatformApi::new(platform_url, std::time::Duration::from_secs(3), logger.clone()).expect("Should make PlatformApi");
+
     match adapter {
         AdapterTypes::Ethereum(adapter) => {
             Application::new(
@@ -115,6 +120,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
                 redis,
                 postgres,
                 campaign_remaining,
+                platform_api,
             )
             .run(socket_addr)
             .await
@@ -127,6 +133,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
                 redis,
                 postgres,
                 campaign_remaining,
+                platform_api,
             )
             .run(socket_addr)
             .await
