+++ sentry/src/routes/units_for_slot.rs
@@ -32,10 +32,7 @@ use primitives::{
 use reqwest::Url;
 use serde::{Deserialize, Serialize};
 use slog::{debug, error, warn, Logger};
-// use url::{form_urlencoded, Url};
 use woothee::{parser::Parser, woothee::VALUE_UNKNOWN};
-
-use platform::PlatformApi;
 
 use crate::{Application, ResponseError};
 
@@ -70,7 +67,6 @@ pub(crate) fn service_unavailable() -> Response<Body> {
 pub async fn post_units_for_slot<C>(
     req: Request<Body>,
     app: &Application<C>,
-    platform: PlatformApi,
     // ipfs: IPFS,
     // caches: Caches<C, AU, AT, AS, E>,
 ) -> Result<Response<Body>, ResponseError>
@@ -96,7 +92,7 @@ where
         alexa_rank: Some(1.0),
     };
 
-    let units = match platform.fetch_units(&request_body.ad_slot.ad_type).await {
+    let units = match app.platform_api.fetch_units(&request_body.ad_slot.ad_type).await {
         Ok(units) => units,
         Err(error) => {
             error!(&logger, "Error fetching AdUnits for AdSlot"; "AdSlot" => ?ad_slot, "error" => ?error);
@@ -109,7 +105,7 @@ where
     let fallback_unit: Option<AdUnit> = match &ad_slot_response.slot.fallback_unit {
         Some(unit_ipfs) => {
             let ipfs = unit_ipfs.parse::<IPFS>()?;
-            let ad_unit_response = match platform.fetch_unit(ipfs.clone()).await {
+            let ad_unit_response = match app.platform_api.fetch_unit(ipfs.clone()).await {
                 Ok(Some(response)) => {
                     debug!(&logger, "Fetched AdUnit"; "AdUnit" => &ipfs);
                     response
@@ -384,163 +380,3 @@ async fn apply_targeting(
     //         .collect()
 }
 
-/// previously fetched from the market (in the supermarket) it should now be fetched from the Platform!
-mod platform {
-    use primitives::{
-        market::{AdSlotResponse, AdUnitResponse, AdUnitsResponse, Campaign, StatusType},
-        util::ApiUrl,
-        AdUnit, IPFS,
-    };
-    use reqwest::{Client, Error, StatusCode};
-    use slog::{info, Logger};
-    use std::{fmt, sync::Arc};
-
-    use crate::Config;
-
-    pub type Result<T> = std::result::Result<T, Error>;
-
-    #[derive(Debug, Clone)]
-    /// The `PlatformApi` is cheap to clone as it already wraps the real client `PlatformApiInner` in an `Arc`
-    pub struct PlatformApi {
-        inner: Arc<PlatformApiInner>,
-    }
-
-    impl PlatformApi {
-        /// The Market url that was is used for communication with the API
-        pub fn url(&self) -> &ApiUrl {
-            &self.inner.platform_url
-        }
-
-        // todo: Instead of associate function, use a builder
-        pub fn new(platform_url: ApiUrl, config: &Config, logger: Logger) -> Result<Self> {
-            Ok(Self {
-                inner: Arc::new(PlatformApiInner::new(platform_url, config, logger)?),
-            })
-        }
-
-        pub async fn fetch_unit(&self, ipfs: IPFS) -> Result<Option<AdUnitResponse>> {
-            self.inner.fetch_unit(ipfs).await
-        }
-
-        pub async fn fetch_units(&self, ad_type: &str) -> Result<Vec<AdUnit>> {
-            self.inner.fetch_units(ad_type).await
-        }
-    }
-
-    /// Should we query All or only certain statuses
-    #[derive(Debug)]
-    pub enum Statuses<'a> {
-        All,
-        Only(&'a [StatusType]),
-    }
-
-    impl fmt::Display for Statuses<'_> {
-        fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-            use Statuses::*;
-
-            match self {
-                All => write!(f, "all"),
-                Only(statuses) => {
-                    let statuses = statuses.iter().map(ToString::to_string).collect::<Vec<_>>();
-
-                    write!(f, "status={}", statuses.join(","))
-                }
-            }
-        }
-    }
-
-    #[derive(Debug, Clone)]
-    struct PlatformApiInner {
-        platform_url: ApiUrl,
-        client: Client,
-        logger: Logger,
-    }
-
-    impl PlatformApiInner {
-        /// The limit of Campaigns per page when fetching
-        /// Limit the value to MAX(500)
-        const MARKET_CAMPAIGNS_LIMIT: u64 = 500;
-        /// The limit of AdUnits per page when fetching
-        /// It should always be > 1
-        const MARKET_AD_UNITS_LIMIT: u64 = 1_000;
-
-        pub fn new(platform_url: ApiUrl, config: &Config, logger: Logger) -> Result<Self> {
-            // @TODO: maybe add timeout?
-            let client = Client::builder()
-                // TODO: Move this config value from Supermarket if needed
-                // .tcp_keepalive(config.market.keep_alive_interval)
-                .cookie_store(true)
-                .build()?;
-
-            Ok(Self {
-                platform_url,
-                client,
-                logger,
-            })
-        }
-
-        pub async fn fetch_unit(&self, ipfs: IPFS) -> Result<Option<AdUnitResponse>> {
-            let url = self
-                .platform_url
-                .join(&format!("units/{}", ipfs))
-                .expect("Wrong Platform Url for /units/{IPFS} endpoint");
-
-            match self.client.get(url).send().await?.error_for_status() {
-                Ok(response) => {
-                    let ad_unit_response = response.json::<AdUnitResponse>().await?;
-
-                    Ok(Some(ad_unit_response))
-                }
-                // if we have a `404 Not Found` error, return None
-                Err(err) if err.status() == Some(StatusCode::NOT_FOUND) => Ok(None),
-                Err(err) => Err(err),
-            }
-        }
-
-        pub async fn fetch_units(&self, ad_type: &str) -> Result<Vec<AdUnit>> {
-            let mut units = Vec::new();
-            let mut skip: u64 = 0;
-            let limit = Self::MARKET_AD_UNITS_LIMIT;
-
-            loop {
-                // if one page fail, simply return the error for now
-                let mut page_results = self.fetch_units_page(ad_type, skip).await?;
-                // get the count before appending the page results to all
-                let count = page_results.len() as u64;
-
-                // append all received units
-                units.append(&mut page_results);
-                // add the number of results we need to skip in the next iteration
-                skip += count;
-
-                // if the Market returns < market fetch limit
-                // we've got all AdSlots from all pages!
-                if count < limit {
-                    // so break out of the loop
-                    break;
-                }
-            }
-
-            Ok(units)
-        }
-
-        /// `skip` - how many records it should skip (pagination)
-        async fn fetch_units_page(&self, ad_type: &str, skip: u64) -> Result<Vec<AdUnit>> {
-            let url = self
-                .platform_url
-                .join(&format!(
-                    "units?limit={}&skip={}&type={}",
-                    Self::MARKET_AD_UNITS_LIMIT,
-                    skip,
-                    ad_type,
-                ))
-                .expect("Wrong Market Url for /units endpoint");
-
-            let response = self.client.get(url).send().await?;
-
-            let ad_units: AdUnitsResponse = response.json().await?;
-
-            Ok(ad_units.0)
-        }
-    }
-}
