+++ test_harness/src/lib.rs
@@ -914,7 +914,7 @@ pub mod run {
             postgres_connection, redis_connection, redis_pool::Manager,
             tests_postgres::setup_test_migrations, CampaignRemaining,
         },
-        Application,
+        Application, platform::PlatformApi,
     };
     use slog::info;
     use subprocess::{Popen, PopenConfig, Redirection};
@@ -950,15 +950,22 @@ pub mod run {
             .await
             .expect("Should flush redis database");
 
+        let logger = new_logger(&validator.sentry_logger_prefix);
         let campaign_remaining = CampaignRemaining::new(redis.clone());
+
+        // todo: Make platform_url configurable! Load from config or pass with env. variable
+    let platform_url = "https://todo-local-platform".parse().expect("Bad ApiUrl, load from Config?");
+    // todo: Make keep_alive_interval configurable!
+    let platform_api = PlatformApi::new(platform_url, std::time::Duration::from_secs(3), logger.clone()).expect("Should make PlatformApi");
 
         let app = Application::new(
             adapter,
             GANACHE_CONFIG.clone(),
-            new_logger(&validator.sentry_logger_prefix),
+            logger,
             redis.clone(),
             postgres.clone(),
             campaign_remaining,
+            platform_api
         );
 
         // Before the tests, make sure to flush the DB from previous run of `sentry` tests
